+++ nym-vpn-core/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix discovery propagation bug (https://github.com/nymtech/nym-vpn-client/pull/4226)
 - Ensure that vpn topology is refreshed periodically when connecting (https://github.com/nymtech/nym-vpn-client/pull/4228)
+- [Android] Bypass local DNS servers (https://github.com/nymtech/nym-vpn-client/pull/4347)
 
 ### Removed
 - Removed credentials mode feature flag from code base (https://github.com/nymtech/nym-vpn-client/pull/4223)